+++ src/main/java/frc/robot/commands/Drive/ArcadeDrive.java
@@ -1,22 +1,21 @@
-package frc.robot.commands;
+package frc.robot.commands.Drive;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.controllers.DriverController;
 import frc.robot.subsystems.DiffDriveSubsystem;
-public class TankDrive extends CommandBase {
-    private final DiffDriveSubsystem diffDriveSubsystem = DiffDriveSubsystem.getInstance();
 
+public class ArcadeDrive extends CommandBase {
+    private final DiffDriveSubsystem diffDriveSubsystem = DiffDriveSubsystem.getInstance();
     private DriverController driver;
 
-    public TankDrive(DriverController myDriver) {
+    public ArcadeDrive(DriverController myDriver) {
         driver = myDriver;
-
         addRequirements(this.diffDriveSubsystem);
     }
 
     @Override
     public void execute() {
-        diffDriveSubsystem.tankDrive(driver.getArcadeFwd(), driver.getArcadeTurn());
+        diffDriveSubsystem.arcadeDrive(driver.getArcadeFwd(), driver.getArcadeTurn());
     }
 
     @Override
@@ -25,4 +24,5 @@ public class TankDrive extends CommandBase {
             diffDriveSubsystem.stop();
         }
     }
+
 }
