+++ src/main/java/frc/robot/commands/Drive/SwitchSpeed.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.Drive;
 
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import frc.robot.RobotState;
